refactor todo retos, ahora es ejercicio-diario
