+++ src/app/app-routing.module.ts
@@ -29,7 +29,7 @@ import { ErrorComponent } from './vistas/error/error.component';
 import { FormAdminComponent } from './vistas/form-admin/form-admin.component';
 import { ListadoDeEjerciciosComponent } from './vistas/listado-de-ejercicios/listado-de-ejercicios.component';
 import { adminGuard } from './core/guardias/admin/admin.guard';
-import { RetosComponent } from './vistas/retos/retos.component';
+import { EjercicioDiarioComponent } from './vistas/ejercicio-diario/ejercicio-diario.component';
 
 const routes: Routes = [
   {
@@ -181,9 +181,9 @@ const routes: Routes = [
         component: FormAdminComponent
       },
       {
-        path: 'retos',
-        component: RetosComponent,
-      }
+        path: 'ejercicio-diario',
+        component: EjercicioDiarioComponent,
+      },
     ]
   },
   {
