+++ src/app/compartido/componentes/carrusel-vertical/carrusel-vertical/carrusel-vertical.component.html
@@ -7,7 +7,7 @@
     <app-boton-trainin type="button" 
              [clase]="'btn-practicar2'" 
              [tipo]="'button'"  
-             ruta="/retos"
+             ruta="/EjercicioDiario"
           ><span style="color:white">
             Ir al Reto
           </span>
