+++ src/app/vistas/ejercicio-diario/ejercicio-diario.component.css
@@ -1,4 +1,4 @@
-.retos-container {
+.ejercicio-diario-container {
   padding: 20px;
   width: 100%;
   margin-top: 65px;
@@ -306,7 +306,7 @@
 
 /* Responsive */
 @media (max-width: 768px) {
-  .retos-container {
+  .ejercicio-diario-container {
     padding: 15px;
   }
   
@@ -332,7 +332,7 @@
     font-size: 80px;
   }
   @media (max-width: 768px) {
-  .retos-container {
+  .ejercicio-diario-container{
     flex-direction: column;
     gap: 30px;
   }
