+++ src/app/vistas/ejercicio-diario/ejercicio-diario.component.cy.ts
@@ -1,0 +1,7 @@
+import {EjercicioDiarioComponent } from './ejercicio-diario.component'
+
+describe('RetosComponent', () => {
+  it('should mount', () => {
+    cy.mount(EjercicioDiarioComponent)
+  })
+})
