--- src/app/vistas/retos/retos.component.cy.ts
@@ -1,7 +1,0 @@
-import { RetosComponent } from './retos.component'
-
-describe('RetosComponent', () => {
-  it('should mount', () => {
-    cy.mount(RetosComponent)
-  })
-})
